requirements.txt
# SkillSync AI - Dependencies
# Python 3.9+ required

# Core Web Framework
streamlit>=1.28.0

# AI/ML
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
openai>=1.0.0

# Data Processing
pandas>=2.0.0

# PDF Processing
PyMuPDF>=1.23.0

# Data Generation
faker>=19.0.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
uuid6>=2024.1.12
